Introducing `account_org_paths` to limit the reachable accounts by OU path (#152)
This PR introduces a new optional account_org_paths to the scanner role. This new variable can be used to enforce that the scanner can only reach for roles in the specified organisational unit paths.

Now that the account_roles is optional and defaults to a wildcard allowing to assume roles from any account, to limit the scope of the scanner role, without losing the convenience of specifying a account wildcard.