Fix Bicep pre-commit hook
The Bicep CLI version in CI is updated periodically. We need to ensure
that we run the pre-commit hook with the same version that was
originally used to build the template, otherwise there will always be
changes and the pre-commit hook will fail.

We can get the exact version used from `.metadata._generator.version` in
the template itself. However, `az bicep install` expects the version to
be passed in a different format: only the first three components and
with a `v` prefix. We use `jq` to convert it to the expected format.