Rework Bicep pre-commit hook (#212)
We need to run this hook not just when the main template changes, but
also whenever any of the template inputs change, including the scripts.

To simplify (and future-proof) things, let's run the hook whenever any
of the files in this directory change (except the README).

I also combined the azure-install and azure-build hooks because they
always need to run together anyway.